SMILE refactoring
- refactored packages and classes from metadb to smile
- updated prop names
