+++ .circleci/config.yml
@@ -10,19 +10,19 @@ orbs:
 jobs:
   run_checkstyle:
     <<: *defaults
-    working_directory: /tmp/repos/cmo-metadb-consistency-checker
+    working_directory: /tmp/repos/smile-consistency-checker
     steps:
       - checkout
       - run:
           name: Pull cmo metadb common code
           command: |
-              export METADB_COMMON_VERSION=$(grep 'cmo_metadb_common\.version' pom.xml | head -1 | sed 's/<cmo_metadb_common\.version>//g' | sed 's|</cmo_metadb_common\.version>||' | tr -d '[:blank:]' | cut -f1) && \
-              export METADB_COMMON_ORG=$(grep 'cmo_metadb_common\.groupId' pom.xml | head -1 | sed 's/<cmo_metadb_common\.groupId>//g' | sed 's|</cmo_metadb_common\.groupId>||' | tr -d '[:blank:]' | cut -d. -f3) && \
+              export SMILE_COMMONS_VERSION=$(grep 'smile_commons\.version' pom.xml | head -1 | sed 's/<smile_commons\.version>//g' | sed 's|</smile_commons\.version>||' | tr -d '[:blank:]' | cut -f1) && \
+              export SMILE_COMMONS_ORG=$(grep 'smile_commons\.groupId' pom.xml | head -1 | sed 's/<smile_commons\.groupId>//g' | sed 's|</smile_commons\.groupId>||' | tr -d '[:blank:]' | cut -d. -f3) && \
               cd .. && \
-              git clone https://github.com/$METADB_COMMON_ORG/cmo-metadb-common.git && \
-              cd cmo-metadb-common && \
+              git clone https://github.com/$SMILE_COMMONS_ORG/smile-commons.git && \
+              cd smile-commons && \
               git fetch --tags && \
-              git checkout ${METADB_COMMON_VERSION}
+              git checkout ${SMILE_COMMONS_VERSION}
       - run:
           name: "Build project..."
           command: mvn clean install
@@ -31,9 +31,9 @@ jobs:
           command: mvn checkstyle:checkstyle
       - run:
           name: "Validating checkstyle plugin report..."
-          command: bash /tmp/repos/cmo-metadb-common/test/scripts/generate_checkstyle_report.sh /tmp/repos/cmo-metadb-consistency-checker
+          command: bash /tmp/repos/smile-commons/test/scripts/generate_checkstyle_report.sh /tmp/repos/smile-consistency-checker
       - store_artifacts:
-          path: /tmp/repos/cmo-metadb-consistency-checker/checkstyle_report_compiled.txt
+          path: /tmp/repos/smile-consistency-checker/checkstyle_report_compiled.txt
           destination: /checkstyle_report.txt
 
 workflows:
