+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -26,7 +26,6 @@ Include specific tasks in the order they need to be done in. Include links to sp
 - [ ] task 3
 
 ---
-
 ## Crossing T's and dotting I's
 
 Please follow these checklists to help prevent any unexpected issues from being introduced by the changes in this pull request. If an item does not apply then indicate so by surrounding the line item with `~~` to strikethrough the text. See [basic writing and formatting syntax](https://docs.github.com/en/github/writing-on-github/getting-started-with-writing-and-formatting-on-github/basic-writing-and-formatting-syntax) for more information.
@@ -44,8 +43,8 @@ Please describe how the workflow and messaging was tested/simulated:
 
 - NATS [local, local docker, dev server, production]
 - Neo4j [local, local docker, dev server, production]
-- MetaDB [local, local docker, dev server, production]
-- Message publishing simulation [nats cli, docker nats cli, metadb publisher tool, other (describe below)]
+- SMILE Server [local, local docker, dev server, production]
+- Message publishing simulation [nats cli, docker nats cli, smile publisher tool, other (describe below)]
 
 Other: [insert details on how messages were published or simulated for testing]
