+++ pom.xml
@@ -1,12 +1,12 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>org.mskcc.cmo.metadb</groupId>
+  <groupId>org.mskcc.smile</groupId>
   <artifactId>master</artifactId>
   <packaging>jar</packaging>
-  <name>MSKCC CMO MetaDB Consistency Checker</name>
+  <name>SMILE Consistency Checker</name>
   <version>0.1.0</version>
   <description>master maven module</description>
-  <url>https://github.mskcc.org/cmo/metadb/</url>
+  <url>https://github.com/mskcc/smile-consistency-checker/</url>
 
   <parent>
     <groupId>org.springframework.boot</groupId>
@@ -35,26 +35,26 @@
     <spring.boot.version>2.3.3.RELEASE</spring.boot.version>
     <slf4j.version>1.7.30</slf4j.version>
     <jackson.version>2.11.2</jackson.version>
-    <!-- metadb messaging -->
-    <cmo_metadb_messaging_java.group>com.github.mskcc</cmo_metadb_messaging_java.group>
-    <cmo_metadb_messaging_java.version>1.2.2.RELEASE</cmo_metadb_messaging_java.version>
-    <!-- metadb common centralized config properties -->
-    <cmo_metadb_common.groupId>com.github.mskcc</cmo_metadb_common.groupId>
-    <cmo_metadb_common.version>1.2.2.RELEASE</cmo_metadb_common.version>
+    <!-- smile messaging java -->
+    <smile_messaging_java.group>com.github.mskcc</smile_messaging_java.group>
+    <smile_messaging_java.version>1.3.1.RELEASE</smile_messaging_java.version>
+    <!-- smile commons centralized config properties -->
+    <smile_commons.groupId>com.github.mskcc</smile_commons.groupId>
+    <smile_commons.version>1.3.1.RELEASE</smile_commons.version>
   </properties>
 
   <dependencies>
-      <!-- cmo messaging library -->
+      <!-- smile messaging library -->
     <dependency>
-      <groupId>${cmo_metadb_messaging_java.group}</groupId>
-      <artifactId>cmo-messaging-java</artifactId>
-      <version>${cmo_metadb_messaging_java.version}</version>
+      <groupId>${smile_messaging_java.group}</groupId>
+      <artifactId>smile-messaging-java</artifactId>
+      <version>${smile_messaging_java.version}</version>
     </dependency>
-    <!-- cmo metadb common -->
+    <!-- smile commons -->
     <dependency>
-      <groupId>${cmo_metadb_common.groupId}</groupId>
-      <artifactId>cmo-metadb-common</artifactId>
-      <version>${cmo_metadb_common.version}</version>
+      <groupId>${smile_commons.groupId}</groupId>
+      <artifactId>smile-commons</artifactId>
+      <version>${smile_commons.version}</version>
     </dependency>
     <!-- testing -->
     <dependency>
@@ -144,9 +144,9 @@
         <version>3.1.1</version>
         <dependencies>
           <dependency>
-            <groupId>${cmo_metadb_common.groupId}</groupId>
-            <artifactId>cmo-metadb-common</artifactId>
-            <version>${cmo_metadb_common.version}</version>
+            <groupId>${smile_commons.groupId}</groupId>
+            <artifactId>smile-commons</artifactId>
+            <version>${smile_commons.version}</version>
           </dependency>
         </dependencies>
         <configuration>
@@ -173,7 +173,7 @@
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-maven-plugin</artifactId>
         <configuration>
-          <mainClass>org.mskcc.cmo.metadb.ConsistencyCheckerApp</mainClass>
+          <mainClass>org.mskcc.smile.ConsistencyCheckerApp</mainClass>
         </configuration>
       </plugin>
     </plugins>
