+++ src/main/java/org/mskcc/smile/ConsistencyCheckerApp.java
@@ -1,17 +1,17 @@
-package org.mskcc.cmo.metadb;
+package org.mskcc.smile;
 
 import java.util.concurrent.CountDownLatch;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.mskcc.cmo.messaging.Gateway;
-import org.mskcc.cmo.metadb.service.MessageHandlingService;
+import org.mskcc.smile.service.MessageHandlingService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication(scanBasePackages = {"org.mskcc.cmo.messaging",
-        "org.mskcc.cmo.common.*", "org.mskcc.cmo.metadb.*"})
+        "org.mskcc.smile.commons.*", "org.mskcc.smile.*"})
 public class ConsistencyCheckerApp implements CommandLineRunner {
     private static final Log LOG = LogFactory.getLog(ConsistencyCheckerApp.class);
